src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/custom/document-register-basic-expected.txt

CONSOLE ERROR: line 21: Uncaught TypeError: document.registerElement is not a function
Testing document.registerElement() basic behaviors.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
